requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
multidict==6.0.5
packaging==24.0
pluggy==1.5.0
prompt-toolkit==3.0.43
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv ==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
requests==2.32.2
simple-websocket==1.0.0
six==1.16.0
tomli==2.0.1
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
